requirements.txt
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.32.1
starlette==0.41.3
faiss-cpu>=1.7.4
pypdf>=4.2.0
numpy>=1.26.4
python-pptx==0.6.23
python-multipart>=0.0.6
tiktoken>=0.5.0